+++ ps_funcs/py/config.py
@@ -12,4 +12,26 @@ class PSConfig (configparser.ConfigParser):
 
 
 
-	pass
+
+
+	def __init__ (self, filename: str):
+		self.filename=filename
+		self.parse()
+		
+	def parse (self, filename: str|None)->bool:
+		if filename is None:
+			if self.filename is None:
+				raise PSException ("Filename should be passed to PSConfig either by constructor or to parse function", colorize=True, fatal=True)
+
+
+	def get (self, path: str|tuple)->str:
+	
+		res = ""
+		if isinstance(path, tuple):
+			res+="tuple got: "+str(path)
+		elif isinstance (path, str):
+			res+="str got: "+str(path)
+		else:
+			res +="wrong type got: "+str(type(path))
+		return "just test as for "+res
+